Flyttar ansvaret for rest-kall og andre aksjonar ut frå panelet og over til provider-klassa, det kjens som ei mykje betre ansvarsfordeling
